backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
langserve[all]==0.3.1
pandas==2.3.0
python-multipart==0.0.20
python-dotenv==1.1.0
langchain-openai==0.3.21
pydantic-settings==2.9.1 